+++ src/main/java/com/foru/freebe/auth/controller/AuthController.java
@@ -10,11 +10,11 @@ import org.springframework.web.bind.annotation.RestController;
 import com.foru.freebe.auth.dto.LoginRequest;
 import com.foru.freebe.auth.dto.LoginResponse;
 import com.foru.freebe.auth.model.KakaoUser;
-import com.foru.freebe.auth.service.AuthService;
+import com.foru.freebe.auth.service.KakaoAuthService;
+import com.foru.freebe.auth.service.KakaoLoginService;
 import com.foru.freebe.common.dto.ResponseBody;
 import com.foru.freebe.jwt.model.JwtTokenModel;
 import com.foru.freebe.jwt.service.JwtService;
-import com.foru.freebe.member.service.MemberService;
 
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
@@ -23,8 +23,28 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class AuthController {
     private final JwtService jwtService;
-    private final AuthService authService;
-    private final MemberService memberService;
+    private final KakaoAuthService kakaoAuthService;
+    private final KakaoLoginService kakaoLoginService;
+
+    @PostMapping("/login")
+    public ResponseEntity<ResponseBody<?>> login(@RequestBody LoginRequest loginRequest) {
+
+        String accessToken = kakaoAuthService.getToken(loginRequest.getCode());
+        KakaoUser kakaoUser = kakaoAuthService.getUserInfo(accessToken);
+
+        LoginResponse loginResponse = kakaoLoginService.findOrRegisterMember(kakaoUser, loginRequest.getRoleType());
+
+        ResponseBody<?> responseBody = ResponseBody.builder()
+            .message(loginResponse.getMessage())
+            .data(loginResponse.getProfileName())
+            .build();
+
+        HttpHeaders headers = jwtService.setTokenHeaders(loginResponse.getToken());
+
+        return ResponseEntity.status(HttpStatus.OK)
+            .headers(headers)
+            .body(responseBody);
+    }
 
     @PostMapping("/reissue")
     public ResponseEntity<Void> reissueToken(HttpServletRequest request) {
@@ -37,25 +57,5 @@ public class AuthController {
         return ResponseEntity.status(HttpStatus.OK.value())
             .headers(headers)
             .body(null);
-    }
-
-    @PostMapping("/login")
-    public ResponseEntity<ResponseBody<?>> login(@RequestBody LoginRequest loginRequest) {
-
-        String accessToken = authService.getToken(loginRequest.getCode());
-        KakaoUser kakaoUser = authService.getUserInfo(accessToken);
-
-        LoginResponse loginResponse = memberService.findOrRegisterMember(kakaoUser, loginRequest.getRoleType());
-
-        ResponseBody<?> responseBody = ResponseBody.builder()
-            .message(loginResponse.getMessage())
-            .data(loginResponse.getProfileName())
-            .build();
-
-        HttpHeaders headers = jwtService.setTokenHeaders(loginResponse.getToken());
-
-        return ResponseEntity.status(HttpStatus.OK)
-            .headers(headers)
-            .body(responseBody);
     }
 }
