+++ src/main/java/com/foru/freebe/auth/service/KakaoLoginService.java
@@ -1,4 +1,4 @@
-package com.foru.freebe.member.service;
+package com.foru.freebe.auth.service;
 
 import org.springframework.stereotype.Service;
 
@@ -8,13 +8,9 @@ import com.foru.freebe.errors.errorcode.CommonErrorCode;
 import com.foru.freebe.errors.exception.RestApiException;
 import com.foru.freebe.jwt.model.JwtTokenModel;
 import com.foru.freebe.jwt.service.JwtService;
-import com.foru.freebe.member.dto.PhotographerJoinRequest;
 import com.foru.freebe.member.entity.Member;
-import com.foru.freebe.member.entity.MemberTermAgreement;
 import com.foru.freebe.member.entity.Role;
 import com.foru.freebe.member.repository.MemberRepository;
-import com.foru.freebe.member.repository.MemberTermAgreementRepository;
-import com.foru.freebe.profile.entity.Profile;
 import com.foru.freebe.profile.service.ProfileService;
 
 import jakarta.transaction.Transactional;
@@ -22,11 +18,10 @@ import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
-public class MemberService {
-    private final ProfileService profileService;
+public class KakaoLoginService {
     private final JwtService jwtService;
+    private final ProfileService profileService;
     private final MemberRepository memberRepository;
-    private final MemberTermAgreementRepository memberTermAgreementRepository;
 
     @Transactional
     public LoginResponse findOrRegisterMember(KakaoUser kakaoUser, Role role) {
@@ -40,25 +35,11 @@ public class MemberService {
 
         JwtTokenModel token = jwtService.generateToken(member.getId());
 
-        // 빌더 시작
         LoginResponse.LoginResponseBuilder builder = LoginResponse.builder();
-
-        // 값 설정
         builder = builder.token(token);
         builder = validateRoleType(builder, member);
 
-        // 최종적으로 build 호출하여 객체 생성
         return builder.build();
-    }
-
-    @Transactional
-    public String joinPhotographer(Member member, PhotographerJoinRequest request) {
-        Member photographer = completePhotographerSignup(member);
-
-        savePhotographerAgreements(photographer, request);
-        Profile profile = profileService.initialProfileSetting(photographer, request.getProfileName());
-
-        return profile.getProfileName();
     }
 
     private LoginResponse.LoginResponseBuilder validateRoleType(LoginResponse.LoginResponseBuilder builder,
@@ -83,20 +64,5 @@ public class MemberService {
             .gender(kakaoUser.getGender())
             .build();
         return memberRepository.save(newMember);
-    }
-
-    private Member completePhotographerSignup(Member member) {
-        member.assignRole(Role.PHOTOGRAPHER);
-        return memberRepository.save(member);
-    }
-
-    private void savePhotographerAgreements(Member member, PhotographerJoinRequest request) {
-        MemberTermAgreement memberTermAgreement = MemberTermAgreement.builder()
-            .member(member)
-            .termsOfServiceAgreement(request.getTermsOfServiceAgreement())
-            .privacyPolicyAgreement(request.getPrivacyPolicyAgreement())
-            .marketingAgreement(request.getMarketingAgreement())
-            .build();
-        memberTermAgreementRepository.save(memberTermAgreement);
     }
 }
